requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
jinja2==3.1.4
python-multipart==0.0.19
itsdangerous==2.2.0
bcrypt==4.2.1
weasyprint==63.1
python-dateutil==2.9.0.post0
